requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
click==8.1.7
colorama==0.4.6
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
joblib==1.4.2
numpy==2.1.3
pandas==2.2.3
psycopg2-binary==2.9.10
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
watchfiles==1.0.0
websockets==14.1
